requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
filelock==3.19.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fsspec==2025.7.0
greenlet==3.2.4
gunicorn==23.0.0
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
safetensors==0.6.2
setuptools==70.2.0
SQLAlchemy==2.0.43
sympy==1.13.3
tokenizers==0.22.0
torch==2.8.0 --index-url https://download.pytorch.org/whl/cpu
torchaudio==2.8.0 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.23.0 --index-url https://download.pytorch.org/whl/cpu
tqdm==4.67.1
transformers==4.56.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3